remove new file
